ЛОМАЮ теперь хотя бы категории работаю, а теги нет

+++ routes/web.php
@@ -6,11 +6,14 @@ use App\Http\Controllers\Admin\Category\EditController;
 use App\Http\Controllers\Admin\Category\ShowController;
 use App\Http\Controllers\Admin\Category\StoreController;
 use App\Http\Controllers\Admin\Category\UpdateController;
-use Illuminate\Support\Facades\Route;
-use App\Http\Controllers;
-use App\Http\Controllers\Main\IndexController;
+use App\Http\Controllers\Admin\Main\IndexController;
 use App\Http\Controllers\Admin\Main\AdminController;
 use App\Http\Controllers\Admin\Category\CategoryController;
+use App\Http\Controllers\Admin\Tag\TagController;
+use Illuminate\Support\Facades\Route;
+use App\Http\Controllers;
+
+
 
 
 /*
@@ -30,9 +33,63 @@ use App\Http\Controllers\Admin\Category\CategoryController;
 //    Route::get('/', AdminController::class);
 //});
 
+ // РАБОЧИЙ РОУТ
+
+//Route::name('main.')->group(function() {
+//    Route::get('/', TagController::class);
+//});
+//
+//
+//
+//
+//Route::group(['namespace' => 'App\Http\Controllers\Admin', 'prefix' => 'admin'], function () {
+//    Route::group(['namespace' => 'Main'], function () {
+//        Route::get('/', [TagController::class])->name('admin.index');
+//    });
+//
+//    Route::group(['namespace' => 'App\Http\Controllers\Category', 'prefix' => 'category'], function () {
+//        Route::get('/', TagController::class)->name('admin.category.index');
+//        Route::get('/create', CreateController::class)->name('admin.category.create');
+//        Route::post('/', StoreController::class)->name('admin.category.store');
+//        Route::get('/{category}', ShowController::class)->name('admin.category.show');
+//        Route::get('/{category}/edit', EditController::class)->name('admin.category.edit');
+//        Route::patch('/{category}', UpdateController::class)->name('admin.category.update');
+//        Route::delete('/{category}', DeleteController::class)->name('admin.category.delete');
+//    });
+//
+//    Route::group(['namespace' => 'App\Http\Controllers\Tag', 'prefix' => 'tags'], function () {
+//        Route::get('/', TagController::class)->name('admin.tag.index');
+//        Route::get('/create', CreateController::class)->name('admin.tag.create');
+//        Route::post('/', StoreController::class)->name('admin.tag.store');
+//        Route::get('/{tag}', ShowController::class)->name('admin.tag.show');
+//        Route::get('/{tag}/edit', EditController::class)->name('admin.tag.edit');
+//        Route::patch('/{tag}', UpdateController::class)->name('admin.tag.update');
+//        Route::delete('/{tag}', DeleteController::class)->name('admin.tag.delete');
+//    });
+//
+//
+//});
+//
+//
+//Route::prefix('admin')->group(function (){
+//    Route::name('main')->group(function() {
+//        Route::get('/', AdminController::class);
+//    });
+//});
+
+// РАБОЧИЙ РОУТ
+
+
+
+
+
+
 Route::name('main.')->group(function() {
     Route::get('/', IndexController::class);
 });
+
+
+
 
 
 Route::group(['namespace' => 'App\Http\Controllers\Admin', 'prefix' => 'admin'], function () {
@@ -40,16 +97,30 @@ Route::group(['namespace' => 'App\Http\Controllers\Admin', 'prefix' => 'admin'],
         Route::get('/', [CategoryController::class])->name('admin.index');
     });
 
-    Route::group(['namespace' => 'Category', 'prefix' => 'categories'], function () {
-        Route::get('/', CategoryController::class)->name('admin.category.index');
-        Route::get('/create', CreateController::class)->name('admin.category.create');
-        Route::post('/', StoreController::class)->name('admin.category.store');
-        Route::get('/{category}', ShowController::class)->name('admin.category.show');
-        Route::get('/{category}/edit', EditController::class)->name('admin.category.edit');
-        Route::patch('/{category}', UpdateController::class)->name('admin.category.update');
-        Route::delete('/{category}', DeleteController::class)->name('admin.category.delete');
+    Route::group(['namespace' => 'Category', 'prefix' => 'category'], function () {
+        Route::get('/', Controllers\Admin\Category\CategoryController::class)->name('admin.category.index');
+        Route::get('/create', Controllers\Admin\Category\CreateController::class)->name('admin.category.create');
+        Route::post('/', Controllers\Admin\Category\StoreController::class)->name('admin.category.store');
+        Route::get('/{category}', Controllers\Admin\Category\ShowController::class)->name('admin.category.show');
+        Route::get('/{category}/edit', Controllers\Admin\Category\EditController::class)->name('admin.category.edit');
+        Route::patch('/{category}', Controllers\Admin\Category\UpdateController::class)->name('admin.category.update');
+        Route::delete('/{category}', Controllers\Admin\Category\DeleteController::class)->name('admin.category.delete');
     });
+
+    Route::group(['namespace' => 'Tag', 'prefix' => 'tags'], function () {
+        Route::get('/', Controllers\Admin\Tag\TagController::class)->name('admin.tag.index');
+        Route::get('/create', Controllers\Admin\Tag\CreateController::class)->name('admin.tag.create');
+        Route::post('/', Controllers\Admin\Tag\StoreController::class)->name('admin.tag.store');
+        Route::get('/{tag}', Controllers\Admin\Tag\ShowController::class)->name('admin.tag.show');
+        Route::get('/{tag}/edit', Controllers\Admin\Tag\EditController::class)->name('admin.tag.edit');
+        Route::patch('/{tag}', Controllers\Admin\Tag\UpdateController::class)->name('admin.tag.update');
+        Route::delete('/{tag}', Controllers\Admin\Tag\DeleteController::class)->name('admin.tag.delete');
+    });
+
+
 });
+
+
 
 Route::prefix('admin')->group(function (){
     Route::name('main')->group(function() {
@@ -58,12 +129,18 @@ Route::prefix('admin')->group(function (){
 });
 
 
+
+
+
+
+
+
 //    Route::name('main')->group(function() {
 //        Route::get('/', AdminController::class);
 //    });
 
  // РАБОТАЕТ
-//    Route::prefix('categories')->group(function () {
+//    Route::prefix('category')->group(function () {
 //        Route::name('category')->group(function () {
 //            Route::get('/', function (){
 //                return 2222222;
@@ -73,20 +150,20 @@ Route::prefix('admin')->group(function (){
 
     // РАБОТАЕТ
 
-//    Route::group(['namespace' => 'Category', 'prefix' => 'categories'], function () {
-//        Route::get('/', CategoryController::class);
+//    Route::group(['namespace' => 'Category', 'prefix' => 'category'], function () {
+//        Route::get('/', TagController::class);
 //    });
    // ЭТО РАБОТАЕТ ВНИЗУ !!!
-//        Route::prefix('categories')->group(function () {
+//        Route::prefix('category')->group(function () {
 //        Route::name('category')->group(function () {
-//            Route::get('/', CategoryController::class)->name('category.index');
+//            Route::get('/', TagController::class)->name('category.index');
 //            });
 //        });
 
 
-//    Route::prefix('categories')->group(function () {
+//    Route::prefix('category')->group(function () {
 //        Route::name('category')->group(function () {
-//            Route::get('/', [App\Http\Controllers\Admin\Category\CategoryController::class])->name('category.index');
+//            Route::get('/', [App\Http\Controllers\Admin\Category\TagController::class])->name('category.index');
 //        });
 //    });
